+++ src/platform/ui_engine/patterns/index.tsx
@@ -11,4 +11,5 @@ export * from "./inline-edit";
 export * from "./pagination";
 export * from "./simple-text-editor";
 export * from "./print-button";
+export * from "./print-format";
 export * from "./vendor-quick-create-dialog";